+++ packages/iov-ethereum/src/ethereumconnection.ts
@@ -74,6 +74,10 @@ export interface EthereumConnectionOptions {
   readonly scraperApiUrl?: string;
   /** List of supported ERC20 tokens */
   readonly erc20Tokens?: ReadonlyMap<TokenTicker, Erc20Options>;
+  /** Time between two polls for block and transaction watching in seconds */
+  readonly blockPollInterval?: number;
+  /** Time between two polls account watching */
+  readonly accountPollInterval?: number;
 }
 
 export class EthereumConnection implements BcpConnection {
@@ -86,6 +90,8 @@ export class EthereumConnection implements BcpConnection {
     return new EthereumConnection(baseUrl, chainId, options);
   }
 
+  private readonly blockPollIntervalMs: number;
+  private readonly accountPollIntervalMs: number;
   private readonly rpcClient: HttpJsonRpcClient;
   private readonly myChainId: ChainId;
   private readonly socket: StreamingSocket | undefined;
@@ -95,6 +101,10 @@ export class EthereumConnection implements BcpConnection {
   private readonly codec: EthereumCodec;
 
   constructor(baseUrl: string, chainId: ChainId, options?: EthereumConnectionOptions) {
+    this.blockPollIntervalMs =
+      options && options.blockPollInterval ? options.blockPollInterval * 1000 : 4_000;
+    this.accountPollIntervalMs =
+      options && options.accountPollInterval ? options.accountPollInterval * 1000 : 5_000;
     this.rpcClient = new HttpJsonRpcClient(baseUrl);
     this.myChainId = chainId;
 
@@ -180,9 +190,6 @@ export class EthereumConnection implements BcpConnection {
 
     const transactionId = Parse.transactionId(transactionResult);
 
-    // 12-15 seconds average block time
-    const pollIntervalMs = 4_000;
-
     let pollInterval: NodeJS.Timeout | undefined;
     const blockInfoPending = new DefaultValueProducer<BlockInfo>(
       {
@@ -203,7 +210,7 @@ export class EthereumConnection implements BcpConnection {
               height: confirmedTransaction.height,
               confirmations: confirmedTransaction.confirmations,
             });
-          }, pollIntervalMs);
+          }, this.blockPollIntervalMs);
         },
         onStop: () => {
           clearInterval(pollInterval!);
@@ -394,7 +401,7 @@ export class EthereumConnection implements BcpConnection {
                   this.getBlockHeader(decodeHexQuantity(blockHeaderJson.params.result.number))
                     .then(blockHeader => listener.next(blockHeader))
                     .catch(error => listener.error(error));
-                }, 3_000);
+                }, this.blockPollIntervalMs);
               }
             } else if (blockHeaderJson.id === subscribeRequestId) {
               if (
@@ -462,7 +469,7 @@ export class EthereumConnection implements BcpConnection {
           }
         };
 
-        setInterval(poll, 5_000);
+        setInterval(poll, this.accountPollIntervalMs);
         await poll();
       },
       stop: () => {
@@ -534,7 +541,7 @@ export class EthereumConnection implements BcpConnection {
           };
 
           await poll();
-          pollInterval = setInterval(poll, 4_000);
+          pollInterval = setInterval(poll, this.blockPollIntervalMs);
         },
         stop: () => {
           if (pollInterval) {
@@ -563,7 +570,7 @@ export class EthereumConnection implements BcpConnection {
             if (searchResult.length > 0) {
               resolve(searchResult[0]);
             } else {
-              await sleep(4_000);
+              await sleep(this.blockPollIntervalMs);
             }
           }
         } catch (error) {
@@ -593,7 +600,7 @@ export class EthereumConnection implements BcpConnection {
           };
 
           await poll();
-          pollInterval = setInterval(poll, 4_000);
+          pollInterval = setInterval(poll, this.blockPollIntervalMs);
         },
         stop: () => {
           if (pollInterval) {